Adapt the path after rename
