+++ src/engine/test/source/server/unixDatagram_test.cpp
@@ -12,7 +12,7 @@
 #include <logging/logging.hpp>
 #include <uvw.hpp>
 
-#include <server/unixDatagram.hpp>
+#include <server/endpoints/unixDatagram.hpp>
 
 using namespace engineserver::endpoint;
